Fixed incosistency when logging in
- Setting loggedIn state to true and updating the db only if Spotify login is successful
- Also moved actions out of folders
