--- sound-bubble/imports/ui/actions/Account/index.js
@@ -1,19 +1,0 @@
-export const getCurrentUser = () => {
-  return (dispatch, getState) => {
-    Meteor.call('getMe', function(err, currentUser) {
-      if (err) {
-        console.log('get current user failed', err);
-      } else {
-        console.log(currentUser);
-        dispatch(getCurrentUserSuccess(currentUser));
-      }
-    });
-  };
-};
-
-const getCurrentUserSuccess = currentUser => {
-  return {
-    type: 'ME',
-    user: currentUser
-  };
-};
